+++ src/foldingProvider.ts
@@ -59,6 +59,10 @@ class MavenLogFoldingRangeProvider implements vscode.FoldingRangeProvider {
             if (downloadingSectionStartIdx === undefined && isDownloadingSectionStart(lineText)) {
                 downloadingSectionStartIdx = lineIdx
             } else if (downloadingSectionStartIdx !== undefined && !isDownloadingSectionLine(lineText)) {
+                if (debugLinesStartIdx !== undefined) {
+                    foldingRanges.push(new vscode.FoldingRange(debugLinesStartIdx,lineIdx-1));
+                    debugLinesStartIdx = undefined
+                }
                 // If there is only a single downloadingProgress section, clear that
                 if (downloadingProgressLinesStartIdx === downloadingSectionStartIdx + 1) {
                     downloadingProgressLinesStartIdx = undefined
@@ -71,6 +75,10 @@ class MavenLogFoldingRangeProvider implements vscode.FoldingRangeProvider {
                 // The first downloadingProgress section within an outer downloading section must start on the second line so it does not interfere with the outer downloading section
                 downloadingProgressLinesStartIdx = Math.max(lineIdx, (downloadingSectionStartIdx || -1)+1);
             } else if (downloadingProgressLinesStartIdx !== undefined && !isDownloadingArtifactSectionLine(lineText)) {
+                if (debugLinesStartIdx !== undefined) {
+                    foldingRanges.push(new vscode.FoldingRange(debugLinesStartIdx,lineIdx-1));
+                    debugLinesStartIdx = undefined
+                }
                 foldingRanges.push(new vscode.FoldingRange(downloadingProgressLinesStartIdx,lineIdx-1));
                 downloadingProgressLinesStartIdx = matchers.downloadingLineRegEx.test(lineText) ? lineIdx : undefined
             }
@@ -144,8 +152,8 @@ function isDownloadingArtifactSectionStart(lineText: string) {
 
 function isDownloadingArtifactSectionLine(lineText: string) {
     return matchers.downloadingProgressLineRegEx.test(lineText)
-        || matchers.whitespaceLineRegEx.test(lineText)
-        || isDebugSectionLine(lineText);
+        || matchers.downloadingDebugLineRegEx.test(lineText)
+        || matchers.whitespaceLineRegEx.test(lineText);
 }
 
 function isDebugSectionStart(lineText: string) {
@@ -154,7 +162,7 @@ function isDebugSectionStart(lineText: string) {
 
 function isDebugSectionLine(lineText: string) {
     return matchers.debugLineRangeRegEx.test(lineText) // Regex implicitly includes debugSectionStart lines, whitespace lines
-        && !matchers.downloadingLineRegEx.test(lineText) 
+        && !matchers.downloadingLineRegEx.test(lineText)
         && !matchers.downloadingProgressLineRegEx.test(lineText);
 }